Rename samples to be samples
